subject average program
