+++ aws/dynamodb.yaml
