+++ src/handlers/subjects.ts
@@ -1,5 +1,6 @@
 //Get
 import { CustomDynamoDB } from "../dynamodb/database";
+import { calculateAvgGrade } from "../helpers/calculations";
 import { HTTP_CODES, USER_TYPES } from "../helpers/constants";
 import responseHelper from "../helpers/responseHelper";
 import { validateToken } from "../helpers/validations";
@@ -14,11 +15,11 @@ export const handler = async(event: any) => {
     if (!tokenData) {
         return responseHelper(HTTP_CODES.FORBIDDEN, "Authentication Failed"); 
     }
-
+    //Validate is Student
     if (tokenData.userType !== USER_TYPES.STUDENT) {
         return responseHelper(HTTP_CODES.UNAUTHORIZED, "You are not authorized to perform this action.");
     }
-    
+    //Get users, subjects and homeworks submited from DB
     const user = await usersDB.getItem(tokenData.userEmail);
     const subjectPromises = user.subjects.map((subjectId) => {
         return subjectsDB.getItem(subjectId)
@@ -27,40 +28,26 @@ export const handler = async(event: any) => {
         return homeworkSubmissionDB.query(subjectId, null, null , "subjectId-Index")
         })
     const subjectsArray = await Promise.all(subjectPromises);
-    const homeworkSubmissionArray = await Promise.all(homeworkSubmissionPromises);
-    
-    const calculateAvgGrade = (submissions) => {
-        if (!submissions) {
-            return null
-        }
-        const gradeArray: Array<any> = submissions.map((hmwkSubmitted) => {
-            return hmwkSubmitted.grade
-        })
-        let sumOfGrades = 0;
-
-        for (let grade of gradeArray){ 
-            sumOfGrades += grade
-        }
-        const totalAvg = sumOfGrades/gradeArray.length
-        return totalAvg
-    }
+    const homeworkSubmissionArray = (await Promise.all(homeworkSubmissionPromises)).flat();
+    console.log('1')
 
     if( !user ){
         return responseHelper(HTTP_CODES.FORBIDDEN, "No data found.")
     }
-
+    console.log('2')
     const subjectInfoResponse = subjectsArray.map((subject) => {
         const filteredSubmissions = homeworkSubmissionArray.filter((hmwksubmission) => {
-            return subject.id === hmwksubmission.subjectId
+            return subject.id === hmwksubmission.subjectId && hmwksubmission.studentEmail === user.email
         })
+    console.log('3')
         return {
-            id: subject.id,
-            teacherEmail: subject.teacherEmail,
-            name: subject.name,
-            description: subject.description,
+            id: subjectInfoResponse.id,
+            teacherEmail: subjectInfoResponse.teacherEmail,
+            name: subjectInfoResponse.name,
+            description: subjectInfoResponse.description,
             grade: calculateAvgGrade(filteredSubmissions)
         }
     })
-
+    console.log('4')
     return responseHelper(HTTP_CODES.SUCCESS, "Success.", subjectInfoResponse)
 }
